Commit: Informatics - 2024-2025 - Practicums - Group 3 - Week 02

+++ Informatics/2024-2025/Practicums/Group 3/Week 02/Solutions/Problem 01.c
@@ -3,11 +3,11 @@
 
 int main() {
     int     a   = 9     ;
-    double  b   = 1.37  ;
+    float   b   = 1.37  ;
     char    c   = 'J'   ;
 
     printf("%d\n",  a);
-    printf("%lf\n", b);
+    printf("%f\n",  b);
     printf("%c\n",  c);
 
     return 0;
